Improve pattern matching and error handling
Improve pattern matching

v0.1.0 converted most of the links in my Evernote exports, but missed
some. Upon inspection, Evernote was inserting additional tags between
the opening anchor tag a and the href, like:

<a style="font-weight: bold;" href="evernote:///">Title</a>

This version of the regex correctly matches Evernote links, even if
additional tags have been inserted into the anchor tag.

Add per-file error handling

el2bl.py v0.1.0 would stop if an error was encountered, such as if the
file already existed in the export directory. Now, the script will
handle errors for each file separately, and if an error is encountered,
el2bl.py will print the error message and move on to the next file.

This commit also improves the messages printed to the terminal, and
improves docstring formatting.