Move exception handling back to `input_enex_path`
Commit 62a293d moved exception handling from `input_enex_path`, the user
input function, to `convert_links`, the reusable function for converting
note links in each file. It is unclear why this change was made, and it
would make more sense to add exception handling to the function handling
user input because a wider variety of exceptions would be handled there.

This commit will move exception handling back to `input_enex_path`.
Now that exception handling has been moved there, a `NotADirectoryError`
will be raised if the user input path is not a directory.