Move output directory creation to `convert_links`
The script creates a directory for converted files. The directory
creation logic was in the user input function `input_enex_path`, but
it makes more sense to put this logic into `convert_links`, the function
that writes the file to the output directory. This commit will move the
directory creation to `convert_links`.